requirements.txt
requests
pandas
beautifulsoup4
lxml
jinja2
Pillow
cairosvg
python-slugify
